+++ src/Adapter/MongoDb/MONGO_DB.mjs
@@ -1,3 +1,0 @@
-export const MONGO_DB_DEFAULT_HOST = "database";
-
-export const MONGO_DB_DEFAULT_PORT = 27_017;