+++ src/Adapter/MongoDb/MongoDb.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {{host?: string, port?: number, user?: string, password?: string, database: string}} MongoDb
- */