+++ src/FLUX_MONGO_DB_CONNECTOR.mjs
@@ -1,0 +1,3 @@
+export const FLUX_MONGO_DB_CONNECTOR_DEFAULT_HOST = "mongo-db";
+
+export const FLUX_MONGO_DB_CONNECTOR_DEFAULT_PORT = 27_017;